(c#) VideoPlayer and FFGraph changes * VideoPlayer: rewrite to fix some egde-case bugs. * decode YUV frames directly in the GPU, this improves performance. *  FFGraph changes:   * improve demuxing by calling avcodec_receive_frame() once per avcodec_send_packet() call.   * rewrite VideoConverter to return raw YUV frames instead of RGB24 if the pixel format is YUV420P.   * update FFmpeg to version 7.1, this also updates the P/Invoke code required to use FFmpeg API. * WGL changes:   * add support to decode and render YUV frames.   * use "gl.activeTexture(gl.TEXTURE0)" before binding texture.   * misc changes to PBO usage (still not implemented). * Texture: add support for chroma planes. * fixes to soundplayer_set_property() and videoplayer_set_property() functions.

+++ src-desktop/kdy-e/engine/sound/soundplayer.cs
@@ -122,7 +122,7 @@ public class SoundPlayer {
                 this.Seek(value);
                 break;
             case VertexProps.MEDIA_PROP_PLAYBACK:
-                switch (value) {
+                switch ((int)value) {
                     case VertexProps.PLAYBACK_PLAY:
                         this.Play();
                         break;
